rename the table component back to compactTable
