+++ packages/apps/dev-wallet/src/App/app.tsx
@@ -1,7 +1,7 @@
-import { MediaContextProvider } from '@/Components/Media/media';
 import { DatabaseProvider } from '@/modules/db/db.provider';
 import { LayoutProvider } from '@/modules/layout/layout.provider.tsx';
 import { WalletProvider } from '@/modules/wallet/wallet.provider';
+import { MediaContextProvider } from '@kadena/kode-ui';
 import { useEffect } from 'react';
 import { PromptProvider } from '../Components/PromptProvider/Prompt';
 import { BetaHeader } from './BetaHeader';
