+++ packages/apps/dev-wallet/src/pages/account/Components/ActivityTable.tsx
@@ -1,8 +1,8 @@
 import { noStyleLinkClass } from '@/Components/Accounts/style.css';
-import { CompactTable } from '@/Components/CompactTable/CompactTable';
 import { usePagination } from '@/Components/CompactTable/usePagination';
 import { IActivity } from '@/modules/activity/activity.repository';
 import { MonoOpenInNew } from '@kadena/kode-icons/system';
+import { DataTable } from '@kadena/kode-ui/patterns';
 import { useMemo } from 'react';
 import { Link } from 'react-router-dom';
 
@@ -36,7 +36,7 @@ export function ActivityTable({ activities }: { activities: IActivity[] }) {
     [activities],
   );
   return (
-    <CompactTable
+    <DataTable
       setPage={handlePageChange}
       pageSize={pageSize}
       totalCount={activities.length}
