+++ packages/apps/explorer/src/components/AccountTransfersTable/AccountTransfersTable.tsx
@@ -4,9 +4,9 @@ import { useQueryContext } from '@/context/queryContext';
 import { usePagination } from '@/hooks/usePagination';
 import { graphqlIdFor } from '@/utils/graphqlIdFor';
 import { Heading, Stack } from '@kadena/kode-ui';
+import { DataTable } from '@kadena/kode-ui/patterns';
 import type { FC } from 'react';
 import React, { useEffect, useState } from 'react';
-import { CompactTable } from '../CompactTable/CompactTable';
 import { FormatAmount } from '../CompactTable/utils/formatAmount';
 import { FormatLink } from '../CompactTable/utils/formatLink';
 import { useToast } from '../Toast/ToastContext/ToastContext';
@@ -70,7 +70,7 @@ export const AccountTransfersTable: FC<{ accountName: string }> = ({
     );
 
   return (
-    <CompactTable
+    <DataTable
       setPage={handlePageChange}
       pageSize={pageSize}
       pageInfo={innerData.node!.transfers.pageInfo}
