+++ packages/apps/explorer/src/components/BlockTransactions/BlockTransactions.tsx
@@ -3,9 +3,9 @@ import { useBlockTransactionsQuery } from '@/__generated__/sdk';
 import { usePagination } from '@/hooks/usePagination';
 import { graphqlIdFor } from '@/utils/graphqlIdFor';
 import { Heading, Stack } from '@kadena/kode-ui';
+import { DataTable } from '@kadena/kode-ui/patterns';
 import type { FC } from 'react';
 import React, { useEffect, useState } from 'react';
-import { CompactTable } from '../CompactTable/CompactTable';
 import { FormatJsonParse, FormatLink } from '../CompactTable/utils/formatLink';
 import { formatMultiStepTx } from '../CompactTable/utils/formatMultiStepTx';
 import { FormatStatus } from '../CompactTable/utils/formatStatus';
@@ -74,7 +74,7 @@ export const BlockTransactions: FC<IProps> = ({ hash }) => {
   }
 
   return (
-    <CompactTable
+    <DataTable
       setPage={handlePageChange}
       pageSize={pageSize}
       pageInfo={innerData.node!.transactions.pageInfo}
