+++ packages/apps/explorer/src/pages/_app.tsx
@@ -1,6 +1,5 @@
 // load global styles from @kadena/kode-ui
 import { Analytics } from '@/components/Analytics/Analytics';
-import { MediaContextProvider } from '@/components/Layout/media';
 import { ToastProvider } from '@/components/Toast/ToastContext/ToastContext';
 import { NetworkContextProvider } from '@/context/networksContext';
 import { QueryContextProvider } from '@/context/queryContext';
@@ -8,7 +7,11 @@ import { SearchContextProvider } from '@/context/searchContext';
 import { useRouter } from '@/hooks/router';
 import { useIphoneInputFix } from '@/hooks/useIphoneInputFix';
 import '@components/globalstyles.css';
-import { RouterProvider, useTheme } from '@kadena/kode-ui';
+import {
+  MediaContextProvider,
+  RouterProvider,
+  useTheme,
+} from '@kadena/kode-ui';
 import '@kadena/kode-ui/global';
 import type { AppProps } from 'next/app';
 import Head from 'next/head';
