+++ packages/libs/kode-ui/src/patterns/CompactTable/CompactTableDesktop/styles.css.ts
