+++ packages/libs/kode-ui/src/patterns/CompactTable/CompactTableMobile/styles.css.ts
