+++ packages/libs/kode-ui/src/patterns/CompactTable/LoadingSkeleton/LoadingPill.tsx
