+++ packages/libs/kode-ui/src/patterns/CompactTable/LoadingSkeleton/ValueLoader/ValueLoader.tsx
