+++ packages/libs/kode-ui/src/patterns/CompactTable/LoadingSkeleton/style.css.ts
