+++ packages/libs/kode-ui/src/patterns/CompactTable/LoadingSkeleton/types.ts
