+++ packages/libs/kode-ui/src/patterns/CompactTable/TablePagination/TablePagination.tsx
