+++ packages/libs/kode-ui/src/patterns/CompactTable/TablePagination/styles.css.ts
