+++ packages/libs/kode-ui/src/patterns/CompactTable/utils/__test__/getItem.test.ts
