+++ packages/libs/kode-ui/src/patterns/CompactTable/utils/formatAmount.tsx
