+++ packages/libs/kode-ui/src/patterns/CompactTable/utils/formatDefault.tsx
