+++ packages/libs/kode-ui/src/patterns/CompactTable/utils/formatLink.tsx
