+++ packages/libs/kode-ui/src/patterns/CompactTable/utils/getItem.ts
