+++ packages/libs/kode-ui/src/patterns/index.ts
@@ -5,3 +5,5 @@ export {
   type ICardContentBlockProps,
   type ICardFooterGroupProps,
 } from './Card';
+
+export { CompactTable, ICompactTableProps } from './CompactTable/CompactTable';
